requirements.txt
aiohttp==3.8.1
aiolinebot==0.4.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
fastapi==0.75.0
flake8==4.0.1
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
line-bot-sdk==2.1.0
MarkupSafe==2.1.0
mccabe==0.6.1
motor==2.5.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pymongo==3.12.3
python-decouple==3.6
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
Werkzeug==2.0.3
yarl==1.7.2
